+++ R/get_project.R
@@ -14,13 +14,20 @@
 #'   brazil_td_chem      \tab Top-down with chemical mechanisms\tab CETESB\tab  .csv and .rds\cr
 #'   brazil_bu_chem      \tab Bottom-up  chemical mechanisms\tab CETESB+tunnel\tab  .rds\cr
 #'   brazil_bu_chem_streets      \tab Bottom-up  chemical mechanisms for streets and MUNICH\tab CETESB+tunnel\tab  .rds\cr
-#'   sebr_cb05co2      \tab Top-down SP, MG and RJ, CB05+CO2\tab CETESB+tunnel\tab  .rds\cr
-#'   amazon2014      \tab Top-down Amazon, Mozart\tab CETESB+tunnel\tab  csv and.rds\cr
-#'   curitiba      \tab Bottom-down +GTFS, RADM2\tab CETESB+tunnel\tab  csv and.rds\cr
+#'   sebr_cb05co2      \tab Top-down SP, MG and RJ\tab CETESB+tunnel\tab  .rds\cr
+#'   amazon2014      \tab Top-down Amazon\tab CETESB+tunnel\tab  csv and.rds\cr
+#'   curitiba      \tab Bottom-down +GTFS\tab CETESB+tunnel\tab  csv and.rds\cr
+#'   masp2020      \tab Bottom-down\tab CETESB+tunnel\tab  csv and.rds\cr
+#'   ecuador_td      \tab Top-down\tab EEA\tab  csv and.rds\cr
+#'   ecuador_td_hot      \tab Top-down\tab EEA\tab  csv and.rds\cr
+#'   ecuador_td_hot_month      \tab Top-down\tab EEA\tab  csv and.rds\cr
+#'   moves      \tab Bottom-up\tab US/EPA MOVES \tab  csv and.rds (requires MOVES >=3.0 on Windows)\cr
 #' }
 #' @param url String, with the URL to download VEIN project
 #' @note default case can be any of "brasil", "brazil", "brazil_bu", "brasil_bu", they are
 #' the same
+#' Projects for Ecuador are in development.
+#' In any case, if you find any error, please, send a pull request in github or gitlab.
 #' @importFrom utils download.file untar
 #' @export
 #' @examples \dontrun{
@@ -37,6 +44,14 @@ get_project <- function(directory,
   if(missing(url)){
   if(case %in% c("brasil", "brazil", "brazil_bu", "brasil_bu")) {
     URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/brazil_bu.tar.gz"
+    tf <- paste0(tempfile(), ".tar.gz")
+    utils::download.file(url = URL,
+                         destfile =  tf)
+    utils::untar(tarfile = tf, exdir = directory)
+    message("Your directory is in ", directory)
+
+  } else if(case == "moves_bu"){
+    URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/moves.tar.gz"
     tf <- paste0(tempfile(), ".tar.gz")
     utils::download.file(url = URL,
                          destfile =  tf)
@@ -83,8 +98,40 @@ get_project <- function(directory,
     utils::untar(tarfile = tf, exdir = directory)
     message("Your directory is in ", directory)
 
+  } else if(case %in% c("masp2020")){
+    URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/MASP_2020.tar.gz"
+    tf <- paste0(tempfile(), ".tar.gz")
+    utils::download.file(url = URL,
+                         destfile =  tf)
+    utils::untar(tarfile = tf, exdir = directory)
+    message("Your directory is in ", directory)
+
   } else if(case %in% c("brazil_td_chem")){
     URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/brazil_td_chem.tar.gz"
+    tf <- paste0(tempfile(), ".tar.gz")
+    utils::download.file(url = URL,
+                         destfile =  tf)
+    utils::untar(tarfile = tf, exdir = directory)
+    message("Your directory is in ", directory)
+
+  } else if(case %in% c("ecuador_td")){
+    URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/ecuador_td.tar.gz"
+    tf <- paste0(tempfile(), ".tar.gz")
+    utils::download.file(url = URL,
+                         destfile =  tf)
+    utils::untar(tarfile = tf, exdir = directory)
+    message("Your directory is in ", directory)
+
+  } else if(case %in% c("ecuador_td_hot")){
+    URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/ecuador_td_hot.tar.gz"
+    tf <- paste0(tempfile(), ".tar.gz")
+    utils::download.file(url = URL,
+                         destfile =  tf)
+    utils::untar(tarfile = tf, exdir = directory)
+    message("Your directory is in ", directory)
+
+  } else if(case %in% c("ecuador_td_hot_month")){
+    URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/ecuador_td_hot_month.tar.gz"
     tf <- paste0(tempfile(), ".tar.gz")
     utils::download.file(url = URL,
                          destfile =  tf)
@@ -100,25 +147,25 @@ get_project <- function(directory,
     message("Your directory is in ", directory)
 
   } else if(case %in% c("curitiba")){
-    URL <- "https://gitlab.com/ibarraespinosa/veinextras/-/raw/master/curitiba_2018.tar.gz"
+    URL <- "https://gitlab.com/ibarraespinosa/veinextras/-/raw/master/curitiba_all/curitiba.tar.gz"
     tf <- paste0(tempfile(), ".tar.gz")
     utils::download.file(url = URL,
                          destfile =  tf)
     utils::untar(tarfile = tf, exdir = directory)
     message("Your directory for vehicular emissions is in ", directory)
 
-    URL <- "https://gitlab.com/ibarraespinosa/veinextras/-/raw/master/gtfs_cur.zip"
+    URL <- "https://gitlab.com/ibarraespinosa/veinextras/-/raw/master/curitiba_all/gtfs_cur.zip"
     tf <- paste0(tempfile(), ".tar.gz")
     utils::download.file(url = URL,
                          destfile =  paste0(directory, "/network/gtfs_cur.zip"))
 
     message("GTFS Curitba is in  ", paste0(directory, "/network"))
 
-    URL <- "https://gitlab.com/ibarraespinosa/veinextras/-/raw/master/curitiba_2018_industrial.tar.gz"
+    URL <- "https://gitlab.com/ibarraespinosa/veinextras/-/raw/master/curitiba_all/curitiba_industrial.tar.gz"
     tf <- paste0(tempfile(), ".tar.gz")
     utils::download.file(url = URL,
                          destfile =  tf)
-    utils::untar(tarfile = tf, exdir = paste0(directory, "_industry"))
+    utils::untar(tarfile = tf, exdir = paste0(directory, "/industry"))
     message("Your directory for industrial emissions is in ", directory)
 
     } else if(case %in% c("sebr_cb05co2")){