+++ NEWS.md
@@ -3,15 +3,29 @@ NEWS
 
 ## TODO
 
-- get_project with Europe, Chinese and MOVES EF
+- get_project with *Europe*, Chinese and MOVES EF
 - Add EF from HBEFA
-- Update ef_ldv* and ef_hdv* to EEA2019
-- Reduce size of sysdata
 
-### vein 0.9.3 (Release date: 2021-06-28)
-- add CBMZ into chem_vein. Based on Yang Zhang Labs data and Carter 2015
+### vein 0.9.4 (5 years!!)
+- add CBMZ into chem_vein. Based on Yang Zhang Labs data and Carter 2015 (Release date: 2021-06-28)
+- add projects `curitiba` and `masp2020`  (Release date: 2021-06-28)
+- fix chem_vein2 for CB4 and CBMZ  (Release date: 2021-07-08)
+- add CB05opt2 into emis_chem2
+- fix fe ETOH on cetesb
+- add add_miles and family add distance
+- add dist argument into Speed to obtain miles/h
+- add function moves_speed to return speed bins used by MOVES
+- changes plots of vein classes, now imports fields::image.plot
+- add moves_exhaust
+- add data(decoder) for MOVES
+- add ef_eea, updated european emission factors
+- add fuel E85 exhaust-evaporative and CNG-exhaust in NMHC speciation
+- add Ecuador
+- For MOVES contributions, added Joao Bazzo
+- reduced size for ef_hdv_speed. Speciation use `speciate`.
+- Deprecated ef_emfac. I never used and it was too heavy.
 
-### vein 0.9.2 (Release date: 2021-06-09)
+### vein 0.9.3 (Release date: 2021-06-09)
 - Fix RCHO in ef_cetesb
 - Change numeric_dc to vector_dv to avoid copying (dotCall64)
 - update gitlab link for veinextras in get_project